general/assembly-code/requirements.txt
keystone-engine
capstone
unicorn